Unsplit US where it should be
