+++ config/splat.us.dra.yaml
@@ -88,9 +88,9 @@ segments:
       - [0x300C, .data, 5F60C] # g_CheatCodes
       - [0x3014, .data, 5F60C] # D_800A3014
       - [0x3134, .data, 628AC] # D_800A3194
-      - [0x31B0, .data, 63ED4] # D_800A3210
-      - [0x3330, .data, 63ED4] # D_800A33A0-D_800A35D0
-      - [0x37D8, .data, 63ED4] # Palette, GFX desc until D_800A3BB8
+      - [0x31B0, .data, 62EB8] # D_800A3210
+      - [0x3330, .data, 62EB8] # D_800A33A0-D_800A35D0
+      - [0x37D8, .data, 62EB8] # Palette, GFX desc until D_800A3BB8
       - [0x3C40, .data, lba_stage]
       - [0x4A00, .data, config_us] # subweapons
       - [0x4B04, .data, config_us] # equipments
@@ -186,7 +186,7 @@ segments:
       - [0x3CC2C, .rodata, 628AC] # func_801028AC
       # - [0x3CC54, .rodata, 628AC] # func_80102D70
       - [0x3CC70, .rodata, 62EB8] # HandleSaveMenu
-      - [0x3CDC0, .rodata, 63ED4] # func_80103ED4
+      - [0x3CDC0, .rodata, 62EB8] # func_80103ED4
       - [0x3CDDC, .rodata, lba_stage]
       - [0x3D188, .rodata, config_us]
       - [0x40D38, .rodata, cd] # func_801080DC
@@ -259,7 +259,6 @@ segments:
       - [0x628AC, c, 628AC]
       - [0x62DEC, c, 62DEC]
       - [0x62EB8, c, 62EB8]
-      - [0x63ED4, c, 63ED4]
       - [0x673C0, c, cd]
       - [0x692E8, c, 692E8]
       - [0x6BF64, c, 6BF64]
